add credits

+++ src/main/java/main/banksystem/containers/Percent.java
@@ -3,9 +3,9 @@ package main.banksystem.containers;
 public enum Percent {
     First(5), Second(10), Third(15), Fourth(20), Fifth(25);
 
-    private int percent;
+    private double percent;
 
-    Percent(int name) {
+    Percent(double name) {
         percent = name;
     }
